% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnip.R
\name{nnip}
\alias{nnip}
\title{nnip}
\usage{
nnip(
  X,
  y,
  test_X,
  valid_X,
  tau,
  hidden_dim1,
  hidden_dim2,
  learning_rate,
  max_deep_iter
)
}
\arguments{
\item{X}{Train predictor data}

\item{y}{Train response data}

\item{test_X}{Test predictor data}

\item{valid_X}{Validation predictor data}

\item{tau}{Target quantiles}

\item{hidden_dim1}{The number of nodes in the first hidden layer}

\item{hidden_dim2}{The number of nodes in the second hidden layer}

\item{learning_rate}{Learning rate in the optimization process}

\item{max_deep_iter}{The number of iterations}
}
\value{
y_predicted, y_test_predicted, y_valid_predited
}
\description{
Estimates non crossing quantile regression with a neural network using interior point method
}
