% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnpg.R
\name{nnqp}
\alias{nnqp}
\title{nnqp}
\usage{
nnqp(
  X,
  y,
  test_X,
  valid_X,
  tau,
  hidden_dim1,
  hidden_dim2,
  learning_rate,
  max_deep_iter,
  penalty = 0.05
)
}
\arguments{
\item{X}{train predictor data}

\item{y}{train response data}

\item{test_X}{test predictor data}

\item{valid_X}{validation predictor data}

\item{tau}{target quantiles}

\item{hidden_dim1}{the number of nodes in the first hidden layer}

\item{hidden_dim2}{the number of nodes in the second hidden layer}

\item{learning_rate}{learning rate in the optimization process}

\item{max_deep_iter}{the number of iterations}

\item{penalty}{the value of tuning parameter for ridge penalty on weights}
}
\value{
y_predicted, y_test_predicted, y_valid_predited
}
\description{
Estimates non crossing quantile regression with a neural network using projected gradient method
}
